+++ src/main/java/com/global/api/entities/enums/ConnectionModes.java
@@ -6,5 +6,6 @@ public enum ConnectionModes {
     SSL_TCP,
     HTTP,
     MEET_IN_THE_CLOUD,
-    DIAMOND_CLOUD
+    DIAMOND_CLOUD,
+    AIDL
 }