+++ src/main/java/com/global/api/terminals/abstractions/IAidlCallback.java
@@ -1,0 +1,5 @@
+package com.global.api.terminals.abstractions;
+
+public interface IAidlCallback {
+    void onResponse(String data);
+}