+++ src/main/java/com/global/api/terminals/abstractions/IAidlService.java
@@ -1,0 +1,7 @@
+package com.global.api.terminals.abstractions;
+
+public interface IAidlService {
+    void onSendAidlMessage(String data, IAidlCallback aidlCallback);
+
+    void onAidlResponse(String data);
+}